+++ src/Backend/MyMusicLibrary.Application/UseCases/Music/GetById/IGetMusicByIdUseCase.cs
@@ -3,5 +3,5 @@
 namespace MyMusicLibrary.Application.UseCases.Music.GetById;
 public interface IGetMusicByIdUseCase
 {
-    Task<ResponseProfileMusicJson> Execute(int id);   
+    Task<ResponseProfileMusicJson> Execute(long id);   
 }
